// frontend/src/app/models/workshops.ts
export class Workshop{
    organizer: string;
    title: string;
    date: string;
    shortDescription: string;
    description: string;
    numberOfLikes: number;
    likes: Array<string>;
    comments: Array<Object>;
    images: Array<string>;
    capacity: number;
    participants: Array<string>;
    waitingList:Array<string>;
    finished: boolean;
    latitude: string;
    longitude: string;
    street: string;
    city: string;
    country:string;
    status: string;
   _id: any;
}